backend/requirements.txt
django
djangorestframework
django-cors-headers
djoser
djangorestframework-simplejwt
psycopg2-binary